Audio Updates
Added background music, a fade in from black, fade in music; all to the mainMenu. LoopAndFadeOutAudio needs some adjustment. For some reason, it doesnt work. I currently have it attached to the audio source, so my guess is that's the problem. But I don't know where it's supposed to go